services/chat/requirements.txt
fastapi
python-dotenv
uvicorn
redis
qdrant_client
google-genai
pydantic
